Move the query to the db module in data.py file
